+++ library/Swagger/Annotations/Scope.php
@@ -1,9 +1,10 @@
 <?php
+
 namespace Swagger\Annotations;
 
 /**
  * @license    http://www.apache.org/licenses/LICENSE-2.0
- *             Copyright [2014] [Robert Allen]
+ *             Copyright [2013] [Robert Allen]
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,41 +22,40 @@ namespace Swagger\Annotations;
  * @category
  * @subpackage
  */
-use Swagger\Annotations\AbstractAnnotation;
+use Swagger\Logger;
 
 /**
+ * Describes an OAuth2 authorization scope.
+ * @link https://github.com/wordnik/swagger-spec/blob/master/versions/1.2.md#516-scope-object
+ *
  * @package
  * @category
  * @subpackage
  *
  * @Annotation
- * @link https://github.com/wordnik/swagger-spec/blob/master/versions/1.2.md#525-response-message-object
  */
-class ResponseMessage extends AbstractAnnotation
+class Scope extends AbstractAnnotation
 {
-    /**
-     * the HTTP response code of the error condition.
-     * @var int
-     */
-    public $code;
 
     /**
-     * The explanation of what will cause the error condition.
+     * The name of the scope.
      * @var string
      */
-    public $message;
-
+    public $scope;
+    
     /**
-     * An optional response class to accompany the response code.
+     * A short description of the scope.
      * @var string
      */
-    public $responseModel;
+    public $description;
 
-    public function __construct(array $values = array())
+	public function validate()
     {
-        parent::__construct($values);
-        if ($this->code !== null) {
-            $this->code = (int) $this->code;
+        if (empty($this->scope)) {
+            Logger::warning('Required field "scope" is missing for "'.$this->identity().'" in '.$this->_context);
+            return false;
         }
-    }
+        return true;
+	}
+
 }